feat: :sparkles: Se agregan los endpoinds de los crup para ta tabla genero, tipo de producto y para la tabla productos
Se agregaron los cruds para 3 tablas, las espesificacion de como se usan esta en el archivo REANDME.md

+++ controller/genero.js
@@ -1,7 +1,6 @@
 import  Express  from "express";
 import conx from './../config/db.js';
 import proxyGenero from "../middleware/proxyGenero.js";
-import e from "express";
 const router = Express();
 
 /* 
@@ -38,8 +37,8 @@ router.post('/',proxyGenero,(req, res)=>{
                 console.log({ "Message": "Error, genero ya existe", "Error": err });
                 res.send({ "Message": "Error, Genero ya existe" });
             }else{
-                console.log({ "Message": "Error al mostrar los generos", "Error": err });
-                res.send({ "Message": "Error al mostrar los generos", "Error": err });
+                console.log({ "Message": "Error al insertar los generos", "Error": err });
+                res.send({ "Message": "Error al insertar los generos", "Error": err });
 
             }
         }else{
@@ -48,20 +47,23 @@ router.post('/',proxyGenero,(req, res)=>{
     })
 })
 
-router.delete('/:id_genero',proxyGenero,(req, res)=>{
+router.delete('/:id_genero',(req, res)=>{
     let params = req.params.id_genero;
-    let query = `DELETE FROM genero WHERE id_genero = ${params}`;
-    console.log(query);
-    conx.query(query,(err,respuesta,fil)=>{
-        if (err) {
-            console.log({ "Message": "Error al eliminar el  genero", "Error": err });
-            res.send({ "Message": "Error al eliminar el  genero", "Error": err });
-        }else{
-            res.send({status:200, Message: "La data se ha eliminado correctamente"});
-        }
-    })
-    
-
+    if (isNaN(Number(params))) {
+        console.log({ "Message": "Error,  Parametros no cumplen con lo espesificado"});
+        res.send({ "Message": "Error, Parametros no cumplen con lo espesificado"});
+    }else{
+        let query = `DELETE FROM genero WHERE id_genero = ${params}`;
+        console.log(query);
+        conx.query(query,(err,respuesta,fil)=>{
+            if (err) {
+                console.log({ "Message": "Error al eliminar el  genero", "Error": err });
+                res.send({ "Message": "Error al eliminar el  genero", "Error": err });
+            }else{
+                res.send({status:200, Message: "La data se ha eliminado correctamente"});
+            }
+        })
+    }
 })
 
 
@@ -74,30 +76,33 @@ Data que le envio
  */
 router.put('/:id_genero',proxyGenero,(req, res)=>{
     let query = `UPDATE genero SET ? WHERE id_genero = ?`;
-    if(!req.body.id_genero && !req.body.nombre_genero){
-        console.log({ "Message": "Error, parametro faltante"});
-        res.send({ "Message": "Error, parametro faltante"});
+    if (isNaN(Number(params))) {
+        console.log({ "Message": "Error,  Parametros no cumplen con lo espesificado"});
+        res.send({ "Message": "Error, Parametros no cumplen con lo espesificado"});
     }else{
-        console.log(query);
-        conx.query(query, [req.body, req.params.id_genero], 
-        (err, respuesta,fil)=>{
-            if (err) {
-                console.log({ "Message": "Error al actualizar el  genero", "Error": err });
-                res.send({ "Message": "Error al actualizar el  genero", "Error": err });
-            }else{
-                console.log(respuesta);
-                if(respuesta.affectedRows === 0){
-                    console.log(respuesta);
-                    res.send({status:200, Message: `el genero con id ${req.params.id_genero} no existe`});
+        if(!req.body.id_genero && !req.body.nombre_genero){
+            console.log({ "Message": "Error, parametro faltante"});
+            res.send({ "Message": "Error, parametro faltante"});
+        }else{
+            console.log(query);
+            conx.query(query, [req.body, req.params.id_genero], 
+            (err, respuesta,fil)=>{
+                if (err) {
+                    console.log({ "Message": "Error al actualizar el  genero", "Error": err });
+                    res.send({ "Message": "Error al actualizar el  genero", "Error": err });
                 }else{
                     console.log(respuesta);
-                    res.send({status:200, Message: "La data se ha actualizado correctamente"});
+                    if(respuesta.affectedRows === 0){
+                        console.log(respuesta);
+                        res.send({status:200, Message: `el genero con id ${req.params.id_genero} no existe`});
+                    }else{
+                        console.log(respuesta);
+                        res.send({status:200, Message: "La data se ha actualizado correctamente"});
+                    }
                 }
-            }
-        })
-    }
-    
-})
-
+            })
+        }
+    }    
+})  
 
 export default router;
